requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==3.0.0
Flask-RESTful==0.3.10

# marshmallow
marshmallow==3.20.1

# env
python-dotenv==1.0.0

# pymongo
pymongo==4.6.3
mongoengine~=0.28.2